patch reduce-scatter test issue
The reduce scatter test is passing in the same global semaphores to reduce scatter for every back to back iteration which is unsafe and can lead to hangs due to race between op iterations using the same global semaphore.

Additionally, reduce scatter currently does not correctly override RT args for global semaphores when the op is rerun with different global semaphores across iterations (to fix first mentioned issue).

For that reason we are patching the test for CI pipeline stability until the reduce scatter RT arg update is resolved. We patch the test by synching after each iteration which removes the race for global semaphore usage between op iterations.