Automatically generate an overload w/o QueueId (#17640)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/10605

### Problem description
This is an experiment aimed to reduce amount of code OP developers
write.

### What's changed
If an OP provides `::invoke` which accepts QueueId as the first
argument, decorator of a registered operation will handle invokes when
`queue_id` is not provided by passing in the `DefaultQueueId`.

While a reduction in code, it comes at a cost of compilation time.
Besides that, the idea of passing queue_id into ops in this way is not
something that I want to support.
This PR is opened to facilitate a conversation.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/13208462636)