#17477: Introduce ND coordinate system for TT-distributed (#17745)
### Ticket
#17477 

### Problem description
Existing mesh infra assumes 2D. This assumption won't hold in the
future.

### What's changed
Introduce a new `SimpleMeshShape` that will gradually replace the
existing `MeshShape`, after which it will be renamed to `MeshShape`.

Introduce `MeshCoordinate`, `MeshCoordinateRange`, and `MeshContainer` -
primitives designed to work with the new ND coordinate system.

`MeshContainer` allows efficient flat representation of various metadata
that matches the mesh shape. Iterators are available to make it easy to
use. `MeshCoordinate` along with strides that are precomputed on
`SimpleMeshShape` allows for an easy point access. The integration with
`MeshBuffer` demonstrates the use case.

Next steps:
* Replace the existing `MeshShape`, `MeshOffset`, and the related
aliases with the new `SimpleMeshShape`, and `MeshCoordinate`.
* No plans to generalize with `CoreCoord`, for now. Cores are
fundamentally in 2D, so a more specialized system can be used for
efficiency. Also it is not desired to make `CoreCoord` to interop with
`MeshCoordinate` - the 2 sets of coordinates mean entirely different
concepts.
* More functionality might be added, as we continue working on
TT-distributed.

### Checklist
- [X] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/13347753550)
- [X] New/Existing tests provide coverage for changes