Current auto-shard heuristic is based on minimising circular buffer size allocation per core, and output tensor buffer size per core.
In this case number of input channels is small, but
number of output channles is large. Since Width
sharding can decouple input and output parallel
config and input is based input channels and output
is based on output channels, output buffer size per
core is small. Problem is currently we ignore the
the input tensor size per core. Width sharding
input parallel core can only use a single core
which means that input tensor and halo output
have to go to a single core and we run out of memory.

Use approx input size per core as factor for
auto-shard heuristic. Ideally we would use halo
output size for this, but using approx input
tensor size is a good enough proxy for now.