#17662: Conv2d fix split reader
In cases where amount of data that needs to be read is uneven,
between first and second reader and in case there are multiple
blocks to be read (case when act_block_h_override is used)
conv2d would fail with pcc issues.

Problem was that offsets for readers between blocks didn't
account for potentially different amount of data being read
by the other reader.