Allow the user to select the version of the docs (#17434)
### Ticket
#17433 

### Problem description

The users would like to see the documentation for several past versions
but we currently only build the documentation for the latest commit on
main.

### What's changed

1. Added a UI selector to be able to switch between version of the
documentation
The versions of the documentation are stored in
https://github.com/tenstorrent/tt-metal/blob/dimitri/test-versioned-docs/docs/published_versions.json

The UI selector makes the HTTP request to get the list of the possible
versions and adds the versions to the list for the user to select.

This required a change to the _layout.html for both `ttnn` and
`tt-metalium`.

2. 
The deployment of Github pages (what we currently use to host
documentation) needs to change to be done from the branch `gh-pages`.
This is needed so that we store previous documentation and only add new
versions to the branch as folders.

The proposed folder structure (e.g.):
```
v0.55.0/ttnn/index.html
v0.54.0/tt-metalium/index.html
```

We will need to change the settings for Github pages to be similar to
the image below:

<img width="1201" alt="Screenshot 2025-01-31 at 1 24 50 PM"
src="https://github.com/user-attachments/assets/c2693721-d8bb-4647-acba-35574d8aa341"
/>


3. We also need to adjust the package workflow to pass the version
number to the docs building workflow and use a different Github actions
to publish the folder.

### Additional Requirements
- [x] Display the version of the docs built next to the library name
- [x] Add a test to verify all links in the Installing.md and Readme.md
- [x] Add a test to make sure that after deploying docs the links to
docs.tenstorrent.com remain accessible


### Checklist
- [x] Prefill the `gh-pages` branch with 3 versions of documentation:
55, 54, and 53
- [x] Switch the population of latest in the workflow
- [x] Flip the setting in the Github pages config (test with @tt-rkim )
- [x] Post commit CI passes 
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] **(For models and ops writers)** Full [new
models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
tests passes
- [ ] New/Existing tests provide coverage for changes