#14080: Preprocess weights for Conv2D on Device (#16750)
### Ticket
#14080

### Problem description
Currently weights preprocessing takes place on the host, on a single
thread. This is slow, especially when there is a large weights matrix,
and Debug mode is enabled.

### What's changed
The weights are loaded to the device in the same format as PyTorch. All
other processing, including permute, padding, etc are done on the
Device.

### Checklist
- [x] Post commit CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13315764885)
- [ ] **(For models and ops writers)** Full [new
models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
tests passes
- [x] New/Existing tests provide coverage for changes