add stream reg use for EDM -> EDM and worker -> EDM (edm side only)
The commit has a mild performance degradation when the 1D fabric kernel is
compiled at opt-level Os (the current compile level on main). However,
when compiled with -O3, the performance is equivalent to slightly
improved for multi-cast but slightly degraded (<2%) for unicast.

Performance measurements are as follows:

Baseline Perf:
- UNICAST perf @ 4k packet: 15.5 GB/s
- MCAST perf @ 4k packet: 12,7 GB/s

Updated Perf:
- UNICAST perf @ 4k packet: 15.2 GB/s
- MCAST perf @ 4k packet: 12.6 GB/s