#0: [skip ci] Fix produce data crash when job is stuck in pending state (#18316)
### Ticket
...

### Problem description
Fix crash in produce data:
https://github.com/tenstorrent/tt-metal/actions/runs/13518418725/job/37810462342#step:10:494
```
get_job_row_from_github_job
    assert github_job["status"] == "completed", f"{github_job_id} is not completed"
AssertionError: 37759949624 is not completed
```

`build-deploy-docs` can sometimes be left in a pending status when
cancelled:

https://github.com/tenstorrent/tt-metal/actions/runs/13514047027/job/37759949624

This causes the produce data flow to trip an assert that only completed
jobs are to be processed.

### What's changed
Downgrade assert to warning and skip the job if its not completed
instead of exiting.

### Checklist
- [x] New/Existing tests provide coverage for changes
Rerun of failed workflow in fix branch
https://github.com/tenstorrent/tt-metal/actions/runs/13530691987