Revert "Revert "Allow the user to select the version of the docs"" (#18349)
Reverts tenstorrent/tt-metal#18291

The second time is a charm. Fixed the issue with the deployment of
assets via Github Pages deploy.
The issue was in .nojekyll file that was needed to instruct Github Pages
deployment to add all files.

Since the .nojekyll file was not present, Jekyll's processing stripped
out all of the folders that started with `_`