Use .ttinsn pseudo (#22)
Now that https://github.com/tenstorrent/tt-metal/pull/17800 has landed,
there's a new `.ttinsn` assembler pseudo. This PR updates the
`ckernel_ops.h` files to make use of that (the same updates as in the
update sfpi headers of that PR). The practical result is that tt insns
are disassembled rather than shown as `.word 0xwhatever`, which is
somewhat more readable.

tt-metal post-commit and blackhole tests pass:
https://github.com/tenstorrent/tt-metal/actions/runs/13522791896
https://github.com/tenstorrent/tt-metal/actions/runs/13522785267