#0: Add TT_METAL_WORKER_RINGBUFFER_SIZE environment variable
This environment variable is a temporary measure that allows users to set the
worker ringbuffer size to see how that affects performance. It will eventually
be replaced with a device parameter.

Also remove common_runtime_address_map.h, since no kernels are actually using
constants from it, and the values it contains should be retrieved in other
ways.