Afuller/try new ci runners (#18663)
### Ticket
#18664

### Problem description
We want to start using the new CI runners. But first we want to test
them out a bit.

### What's changed
Scheduled a job outside of APC to run on a new CI runner.
The same task when run inside APC will continue to run on the existing
runners to minimize potential disruption.
Checked the syntax by having a run that targeted each type:
https://github.com/tenstorrent/tt-metal/actions/runs/13666458620