Minimal All Reduce for Llama shapes (#18217)
### Ticket
- #18218

### Problem description
A minimal All Reduce is required for the low-latency use case for Llama
on TG.

### What's changed
This PR merges in an All Reduce Async that is based on a naive algorithm
- it performs an All Gather, followed by a local reduction.

#### Changes
- Added perf regression checking for llama ccls
- new all reduce async implementation
- Added new function `run_device_perf_detailed` that gets the min, max,
avg, and std for a run

#### Supports:
- Multi-link, single-worker per link
- Resharding
- Constraint: the output number of tiles should be divisible by num
links and output num shards. Ie, each link worker is responsible for
full output shards. Support for this can be added in the future, however
it is not a priority at the moment
- Input tensors that consist of shard padding

### TODO Checklist
- [x] Update validation code
- [x] Add perf regression checking in the new all reduce test
- [x]  Separate subtract CoreRangeSet API into a different PR

### Checklist
- [ ] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI passes
- [x] [TG
Nightly](https://github.com/tenstorrent/tt-metal/actions/runs/13688165179/job/38276633083)
passes

---------

Co-authored-by: kpaigwar <kpaigwar@tenstorrent.com>
Co-authored-by: yugaoTT <yugao@tenstorrent.com>