Remove scheduled cron job from build-and-test-wheels
This workflow was originally intended to test building wheels with and without build-artifact, which should be able to be skipped on 20.04/22.04 GH hosted machines according to @tt-rkim. However, this caused failures when build-artifacts was skipped. In practice we only build wheels from build-artifacts.
Leaving this workflow available to be manually triggered with and without build-artifact but removing the always failing cron job.