#0: Improve test_pgm_dispatch
In particular add new tests that seemed helpful when doing a cost model
calculation (such as having CBs covering portions of the workers, and CRTAS),
as well as a flag to add a lot of random tests.  Also add a flag to report
benchmark parameters in the JSON to make it possible for later scripts to do
further analysis.

Add the current clock frequency to the output so we can detect throttling more
easily.

Also do CloseDevice on some early returns, so later tests that attempt to use
the same device don't run into problems.