[skip ci] Run Smoke tests in PR Gate (#18756)
### Ticket
#17851

### Problem description
No tests are run automatically in PRs
We have the beginning of a 'smoke' test, but are infra constrained

### What's changed
Light up the Smoke tests in the PR Gate, but only where we can target
the new runners which are not (yet) saturated.
We'll monitor to see if any adjustments are needed for workflow
latencies.

### Checklist
- [x] Made sure All Post Commit can [still run Smoke as
before](https://github.com/tenstorrent/tt-metal/actions/runs/13710916570/job/38347445388).