Added mesh support to global circular buffer (#18772)
### Ticket

### Problem description
I'm working on using MeshDevice everywhere in TT-NN, and it runs into an
issue of GlobalCircularBuffer not supporting MeshDevice/MeshBuffer, this
manifests in failing `prefetcher.py` test.

### What's changed
Added support for MeshDevice/MeshBuffer in GlobalCircularBuffer. This,
together with removing TT-NN's `DeviceGlobalCircularBuffer` and
`MultiDeviceGlobalCircularBuffer` resolves the failure. The full change
is shown in this
[commit](https://github.com/tenstorrent/tt-metal/pull/18470/commits/041d3638c6afe6dcfa700f620b6dd85491466fb0)

### Checklist
- [ ] [All post commit CI
passes](https://github.com/tenstorrent/tt-metal/actions/runs/13718502710)
- [x] New/Existing tests provide coverage for changes