[skip ci] Add a Workflow Status action (#18829)
### Ticket
None

### Problem description
GH is being GH

### What's changed
Create an Action that can collect the status of all the jobs in a
workflow so it can report the status of the workflow overall.
Since we'll need to reference this not as a local path, this needs to
land on main before we can start using it.