[skip ci] A Merge Gate workflow (#18896)
### Ticket
Setting the stage for #6857

### Problem description
Failures slip on to `main` with the current flow of
very-expensive-workflow-so-sometimes-not-run on PRs.
Some light tests are beginning to be run by force in PRs. However some
breakages require jobs that are too expensive to run repeatedly on PRs
(case in point: Clang Tidy which has broken several times over the past
7 days).

### What's changed
Defining a Merge Gate that we can then use as we stand up Merge Queues.