Simplify autoformat facilities in preparation for removal (#17896)
### Ticket
https://github.com/tenstorrent/tt-metal/issues/17276

### Problem description
Autoformat is a convoluted mechanism which we want to remove.
But because its convoluted its hard to make in one step.
Want to clean it up a bit as I am trying to better understand how it
works.

### What's changed
Added comments, removed things which clearly should not be needed.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/13734432025)