#18726: Fabric initialization infrastructure for fast dispatch (#18723)
### Ticket
N/A

### Problem description
Dispatch initialization needs to support tt-fabric

### What's changed
- Updated `DispatchMemMap` to reserve L1 space for the Fabric client
interface struct
- Provide the necessary upstream and downstream chip+mesh ID for
Prefetch and Dispatch to use tt-fabric APIs
- Split host and device device initialization in DevicePool to support
the use case below.
- Split static args and dependent args generation in `FDKernel`.
Previously, static + dependent args are initialized on 1 device at a
time. But with this change, it's possible for all static args to be
initialized at once. This lets you put upstream and downstream nodes on
a different device
- This is needed because on Fabric, Prefetch H node (device 0) is
"directly" connected to a Prefetch D (device 1) downstream.
- There is a type conflict when including the fabric host interface on
the host code. This issue is fixed with

https://github.com/tenstorrent/tt-metal/commit/080a2512d2367168df91b4b3119cf3bb75266eca
but I put everything into a namespace anyways to prevent this issue in
the future.
- New intermediary node called `FabricRouterVC` in topology to replace
the tunneling nodes. This node is used indicate that two `FDKernel`
nodes will communicate using fabric. It provides a fabric router
obtained from the `control plane` for the upstream and downstream nodes.

~~~

Additional compile args to be passed into `cq_prefetch` and
`cq_dispatch`

1. `downstream_mesh_id`
2. `downstream_chip_id`
3. `upstream_mesh_id`
4. `upstream_chip_id`
5. `fabric_router_noc_xy`

~~~

### Checklist
APC
https://github.com/tenstorrent/tt-metal/actions/runs/13735730315
T3K
https://github.com/tenstorrent/tt-metal/actions/runs/13735730327
https://github.com/tenstorrent/tt-metal/actions/runs/13774926961
TGG
https://github.com/tenstorrent/tt-metal/actions/runs/13735730706
https://github.com/tenstorrent/tt-metal/actions/runs/13774927277
TG
https://github.com/tenstorrent/tt-metal/actions/runs/13735730523
SingleCard TTNN
https://github.com/tenstorrent/tt-metal/actions/runs/13735730757

---------

Co-authored-by: Austin Ho <aho@tenstorrent.com>
Co-authored-by: Andrew Fuller <afuller@tenstorrent.com>