Preprocess Conv2D weights on Device (#18272)
### Ticket
#14080

### Problem description
Currently weights preprocessing takes place on the host, on a single
thread. This is slow, especially when there is a large weights matrix,
and Debug mode is enabled.

### What's changed
The weights are loaded to the device in the same format as PyTorch. All
other processing, including permute, padding, etc are done on the
Device.

Added two new arguments to Conv2dConfig.
`preprocess_weights_on_device` is a boolean, that if set to true does
the weights preprocessing on the device.
If it's false, preprocessing is first done on the host, and then the
weights are moved to the device.

`always_preprocess_weights` is a boolean, that if set to false disables
preprocessing if the input weights/bias are already on device. If it's
set to true, it will attempt preprocessing regardless of the weights
being on host or device.

Setting to true, and passing preprocessed weights will cause an error. 

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13658485424)
- [x] Blackhole Post commit
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13615764580)
- [x] Nightly tt-metal L2
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13615772023)
- [x] Single Card Demo Tests
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13531412498)
- [x] [Model
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-models.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13625845899)
- [x] [Device performance
regression](https://github.com/tenstorrent/tt-metal/actions/workflows/perf-device-models.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13615768608)
- [ ] Nightly L2
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13515365960)
- [ ] **(For models and ops writers)** Full [new models
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml)
CI
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13780743231)
(if applicable)
- [ ] New/Existing tests provide coverage for changes