Fix conv2d in NO_DISPATCH mode.
Within graph capture NO_DISPATCH mode, the conv2d
is not going to allocate the output tensor.
Conv2d op has logic to calculate memory
footprint in advance, and after op execution it
asserts that this memory was calculated correctly,
based on real usage obtained from allocaator.

In graph capture NO_DISPATCH mode, this checks
breaks, because the output tensor is not allocated.
Currently op can't inffer that it's being executed
in this mode.

As a stop gap solution, op now assumes this mode
of execution if the amount of allocated memory
returned by allocator is zero.