#0: [skip ci] Increase target perf of LLM and remove CCL in TG model perf pipeline (#19031)
### Ticket
N/A

### Problem description
TG model perf pipelines
https://github.com/tenstorrent/tt-metal/actions/workflows/tg-model-perf-tests.yaml
have been failing due to failing LLM and CCL tests.

### What's changed
Checked with @SeanNijjar and @ntarafdar that the CCL perf test can be
removed.
Checked with @cglagovichTT that LLM test should have increased perf
target so it passes but we still check for hangs with the pipeline.

### Checklist
Checked that all tests for TG perf pipeline passes on branch:
https://github.com/tenstorrent/tt-metal/actions/runs/13823649126