#19127: Allow merging CRTA writes with CB/SEM writes.
If we have multiple kernels using CRTAs on the same core, the current code
requires a multicast per kernel, which is inefficient.

Instead, in those cases we can try to merge those writes to a single core
together, as well as with CBs and semaphores. This works better in most cases,
but does require duplicating the CRTA data per kernel-group, which also means
we sometimes need to copy the CRTA data into the command sequence every time we
issue the commands.