[skip ci] Light up the TG Quick test (#19135)
### Ticket
#18419 

### Problem description
Galaxy gets broken because nobody runs it before merging.
Nobody runs it because it's off the beaten path and takes too long and
the resources are scarce.

### What's changed
A fast test that doesn't hog resources placed right on the road well
traveled.

NOTE: We'll monitor for any delays in APC, but for now FAFO'ing it. Easy
to adjust if we need to.

### Checklist
- [x] TG Choose Your Own Pipeline
[passes](https://github.com/tenstorrent/tt-metal/actions/runs/13841186994)