Add PRNG to BH dropout (#43)
**Ticket**
https://github.com/tenstorrent/tt-metal/issues/7736

This PR copies Wormhole's PRNG for dropout into blackhole

**Notes**
First, test to see if `ckernel::wait` works on blackhole, otherwise
fallback to the commented code inside `init_prng_seed`. See
https://github.com/tenstorrent/tt-metal/issues/10868

Second, I've created a tt-metal branch with the complete changes for the
dropout ckernel headers that are in
`tt_metal/hw/ckernels/blackhole/metal/llk_api/llk_sfpu`, and routes the
tt_llk submodule to this PR's branch.

You can use that fork to test as if the full change is applied for both
repos:
https://github.com/JushBJJ/tt-metal/tree/bh-dropout