Migrate CCLs to the new style create_at infra (#19455)
### Ticket
N/A

### Problem description
CCLs need an extension point `create_at` to be able to parametrize
program based on the device coordinate.

### What's changed
Existing CCLs create `struct` per device, which is no longer be the case
with mesh dispatch infra. The per-device parameters should be moved to
program creation time and be provided dynamically.

[X]
[APC](https://github.com/tenstorrent/tt-metal/actions/runs/14001575314)