Enable bfp4 kv_cache for tt_transformers (#19493)
### Ticket
Link to Github Issue

### Problem description
There are TT_FATAL checks in kernel implementations that bans the use of
`bfp4` as dtype for kv_cache. There is no technical reason for this ban
(because `bfp8` is also block float and it is supported). The reason
behind this ban is probably due to the low accuracy in the ML model.

### What's changed
Enable the use of `bfp4` in the relevant kernel implementations to allow
for `bfp4` kv_cache. This is to unblock tradeoff studies by @tlevinTT.

Thanks, @cglagovich, for suggesting to add two select unit test cases
for `bfp4` to the changed ops.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/13998647341)
  - all passed except for one timeout
- [x] [single card nightly
test](https://github.com/tenstorrent/tt-metal/actions/runs/13998655086)
  - all ✅