#18650 Fix sharded unary op writing out of bounds
In case unary op has sharded input, in row major
layout, and shard width is not a multiple of 16
op will be writing out of bounds.

Fix this by removing wrong alignment of shard
width to be multiple of 16 in unary_sharded_program_factory.

This issue doesn't get exposed in unit tests since
output of the op is fine, it's just writing out of
bounds.

This was a root cause of data corruption in SD
model on blackhole. Issue is not blackhole specific,
it was just unfortunate layout in memory.

This issue was also visible as a warning that
was emitted by the op that CB size exceeds the
the buffer size.

# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# Date:      Mon Mar 24 21:50:35 2025 +0000
#
# On branch pjosipovic/fix_silu
# Your branch and 'origin/pjosipovic/fix_silu' have diverged,
# and have 1 and 1 different commits each, respectively.
#   (use "git pull" to merge the remote branch into yours)
#
# Changes to be committed:
#	modified:   ttnn/cpp/ttnn/operations/eltwise/unary/device/unary_sharded_program_factory.cpp
#