Enable assert on device perf by default
We noticed last week that ResNet50 device perf
was returning "inf" and CI wasn't failing.
ResNet got fixed, but we more worrying issue is
that CI didn't catch this. It turns out that
check_device_perf() doesn't assert by default and
only some models opt-in to assert mode.

This PR changes the default to assert on device perf.

Target for BERT perf is adjusted to make the CI
green.